pyproject.toml
[project]
name = "pythonproject12"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.2.0",
    "autogen-agentchat>=0.7.5",
    "autogen-ext>=0.2.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tavily>=1.1.0",
    "tiktoken>=0.12.0",
    "yfinance>=0.2.66",
]
